clear_dats.py keeps last set of .dats for checkpointing (assumes file name is of the form abcNUMBER_xyz.dat)

+++ clear_dats.py
@@ -13,6 +13,8 @@ Script for clearing intermediate .dat files from immediate sub directories.
 Also takes a regex command line argument to filter sub directories.
 '''
 
+regx = re.compile('([0-9]+)_')
+
 def get_immediate_subdirectories(direc):
     return [name for name in os.listdir(direc) if os.path.isdir(os.path.join(direc, name))]
 
@@ -26,5 +28,14 @@ for direc in get_immediate_subdirectories('.'):
     
     flist = glob.glob(direc+'/*.dat')
 
+    #keep last files for checkpointing
+    nums = []
     for f in flist:
-        os.remove(f)
+        nums.append( int(regx.findall(f)[-1]) )
+
+    maxnum = max(nums)
+    print 'Keeping output %d' % maxnum
+
+    for f,num in zip(flist,nums):
+        if num!=maxnum:
+            os.remove(f)
